Reorganizing library files for thermo package
moved the IdealGasPackage struct into its own file and made it into a
submodule of the thermodynamics module. In the future, as more packages
are implemented, one simply needs to create another sub-module when
implementing a new thermodynamics package
